RedditDataEngineering/ProjectDirectory/requirements.txt
apache-airflow==2.8.0
psycopg2-binary==2.9.3
praw==7.7.1
pandas==1.5.3
boto3==1.34.84 